quest pager fix (trunk only)
     A priest quest message that was supposed to say "brotherhood" or
"sisterhood" said "itood" instead.  Text "%shood" used the post-3.4.3 'h'
modifier to substitute a pronoun in place of the %s value.  That's only
a valid modifier when it follows %d (deity), %l (leader), %n (nemesis),
or %o (artifact).  Change the substitution routine to leave it as an 'h'
when it follows anything else.  [No fixes entry needed.]